Rename ops.js to format-feature.js update tests & add a few
