+++ lib/geocoder/format-features.js
@@ -1,10 +1,12 @@
 'use strict';
-module.exports.toFeature = toFeature;
-
 const closestLang = require('../text-processing/closest-lang');
 const featureMatchesLanguage = require('./filter-sources').featureMatchesLanguage;
 const termops = require('../text-processing/termops');
 const bbox = require('../util/bbox');
+const filter = require('./filter-sources');
+
+module.exports.toFeature = toFeature;
+module.exports.toFeatures = toFeatures;
 
 /**
  * getFormatString - get a `place_name` template string from a format object
@@ -39,6 +41,7 @@ function getFormatString(context, format, language) {
  * @param {String} language code to select language/formatting from context array. Undefined defaults to `carmen:text`
  * @param {String} languageMode option that can be set to 'strict' filter results strictly to a specific language. Undefined defaults null.
  * @param {Boolean} matched If set, use matched text in place name
+ * @param {Carmen} geocoder A carmen object used to generate the results
  * @return {String} formatted place_name string
  */
 function getPlaceName(context, formatString, language, languageMode, matched, geocoder) {
@@ -116,6 +119,25 @@ function getPlaceName(context, formatString, language, languageMode, matched, ge
     }
 
     return place_name;
+}
+
+/**
+ * Build a list of index types from a format string
+ * @param {object} index - carmen index
+ * @returns {Set|boolean} set of index types or false
+ */
+function typesFromIndex(index) {
+    if (index.geocoder_format && index.geocoder_format.default) {
+        const p = /(?:[^{]*)\{([^.]+)\.[^}]+\}/y;
+        const types = new Set();
+        let m;
+        // eslint-disable-next-line no-cond-assign
+        while (m = p.exec(index.geocoder_format.default)) types.add(m[1]);
+        return types;
+    }
+    else {
+        return false;
+    }
 }
 
 /**
@@ -240,6 +262,108 @@ function toFeature(context, format, languages, languageMode, debug, geocoder, cl
 }
 
 /**
+ * Takes an array of contexts and returns geojson features, recording
+ * used indexes along the way. Any of our `raw contexts => final output`
+ * logic will ideally end up here.
+ *
+ * @param {Carmen} geocoder A carmen object used to generate the results
+ * @param {array} contexts A contexts array of results to convert to features
+ * @param {object} options A geocoder query options object
+ * @return {object} Object with GeoJSON features on the 'features' property.
+ */
+function toFeatures(geocoder, contexts, options) {
+    let features = [];
+    const indexes = {};
+    const by_place_name = {};
+
+    for (let i = 0; i < contexts.length; i++) {
+        const context = contexts[i];
+        const index = geocoder.indexes[context[0].properties['carmen:index']];
+        if (!filter.featureAllowed(index, context[0], options)) continue;
+
+        // Convert this context array to a feature
+        const feature = toFeature(context, index.geocoder_format, options.language, options.languageMode, options.debug, geocoder, options.clipBBox, options.routing);
+
+        if (!options.allow_dupes) {
+            const types = typesFromIndex(index);
+
+            // Dedupe by place name, prefer feature with non-omitted, non-interpolated geom
+            const keys = [feature.place_name];
+            if (context[0].properties['carmen:spatialmatch'] !== undefined && context[0].properties['carmen:address'] !== undefined) {
+                keys.push(uniqueAddressId(geocoder, types, context));
+            }
+
+            const previous = keys.reduce((m, v) => (m || by_place_name[v] || m), false);
+            if (previous) {
+                if (features[previous.index].address && !feature.address) {
+                    continue; // Don't allow street fallbacks to replace omitted or interpolated addresses
+                }
+                else if (features[previous.index].geometry.omitted && !feature.geometry.omitted) {
+                    features[previous.index] = feature;
+                }
+                else if (features[previous.index].geometry.interpolated && !feature.geometry.interpolated) {
+                    features[previous.index] = feature;
+                }
+                continue;
+            }
+
+            for (const k of keys) {
+                by_place_name[k] = { feature: feature, index: features.length };
+            }
+        }
+        features.push(feature);
+
+        // Record index usage for this feature
+        if (options.indexes) for (let k = 0; k < context.length; k++) {
+            indexes[context[k].properties['carmen:index']] = true;
+        }
+    }
+
+    // Re-sort array if we've changed composition
+    if (features.length !== contexts.length) {
+        features.sort((a, b) => b.relevance - a.relevance);
+    }
+
+    if (options.limit) features = features.slice(0, options.limit);
+
+    const data = { features: features };
+    if (options.indexes) data.indexes = Object.keys(indexes);
+    return data;
+}
+
+/**
+ * Generate an unique Id for an address results based on the cover text and
+ * the context array.
+ *
+ * @param {object} geocoder - carmen object
+ * @param {boolean|Set} types - set of types
+ * @param {Array<object>} context - context array
+ * @returns {string} unique result id
+ */
+function uniqueAddressId(geocoder, types, context) {
+    const coverText = new Map();
+    const covers = context[0].properties['carmen:spatialmatch'].covers;
+    let text = '';
+    for (let i = 0; i < covers.length; i++) {
+        text += ` ${covers[i].text}`;
+    }
+    coverText.set(context[0].properties['carmen:index'], text);
+
+    for (let j = 1; j < context.length; j++) {
+        const indexName = context[j].properties['carmen:index'];
+        // If we already have an string to represent this context, skip it.
+        if (coverText.has(indexName)) continue;
+        // If we got a list of acceptable types, respect it.
+        if (!types || types.has(geocoder.indexes[indexName].type)) {
+            coverText.set(indexName, context[j].properties['carmen:extid']);
+        }
+    }
+
+    return '_' + Array.from(coverText.values()).join(':');
+}
+
+
+/**
  * Attempt to locate text that closely matches the user's query
  * @param {object} item - GeoJSON feature
  * @param {object} geocoder - carmen object
@@ -345,14 +469,14 @@ function getMatchingText(item, geocoder, requestedLanguage) {
     }
 }
 
-/*
-* getIntersectionStreetName - helper function to get the street to add to the cross street
-* this is for Street B in Street A and Street B
-* the carmen:text field has a number of synonyms, we need to pick out the one similar to what the user typed
-* @param {Object} geocoder Geocoder object
-* @param {object} properties A feature properties object
-* @return String returns the correct carmen:text synonym based on the input query
-*/
+/**
+ * getIntersectionStreetName - helper function to get the street to add to the cross street
+ * this is for Street B in Street A and Street B
+ * the carmen:text field has a number of synonyms, we need to pick out the one similar to what the user typed
+ * @param {Object} geocoder Geocoder object
+ * @param {object} properties A feature properties object
+ * @return {string} returns the correct carmen:text synonym based on the input query
+ */
 function getIntersectionStreetName(geocoder, properties) {
     if (properties['carmen:intersection'] && properties['carmen:query_text'] !== undefined) {
         let featureSynonymMatch = ' ';
