+++ lib/geocoder/geocode.js
@@ -1,5 +1,5 @@
 'use strict';
-const ops = require('./ops'),
+const format = require('./format-features'),
     phrasematch = require('./phrasematch'),
     context = require('./context'),
     termops = require('../text-processing/termops'),
@@ -10,8 +10,7 @@ const ops = require('./ops'),
     token = require('../text-processing/token'),
     cl = require('../text-processing/closest-lang'),
     constants = require('../constants'),
-    errcode = require('err-code'),
-    filter = require('./filter-sources');
+    errcode = require('err-code');
 
 /**
  * Main interface for querying an index and returning ranked results.
@@ -192,7 +191,7 @@ function idGeocode(geocoder, asId, options, callback) {
         };
         for (let i = 0; i < features.length; i++) {
             if (!features[i]) continue;
-            const f = ops.toFeature([features[i]]);
+            const f = format.toFeature([features[i]]);
             f.relevance = 1;
             resultIndexes[features[i].properties['carmen:index']] = true;
             result.features.push(f);
@@ -312,7 +311,7 @@ function reverseGeocode(geocoder, tokenized, options, callback) {
      */
     function finalize(err, contexts) {
         try {
-            Object.assign(queryData, toFeatures(geocoder, contexts, options));
+            Object.assign(queryData, format.toFeatures(geocoder, contexts, options));
         } catch (err) {
             return callback(err);
         }
@@ -463,7 +462,7 @@ function forwardGeocode(geocoder, text, options, callback) {
         }
 
         try {
-            Object.assign(queryData, toFeatures(geocoder, contexts, options));
+            Object.assign(queryData, format.toFeatures(geocoder, contexts, options));
         } catch (err) {
             return callback(err);
         }
@@ -477,125 +476,5 @@ function forwardGeocode(geocoder, text, options, callback) {
         if (options.debug) queryData.debug = options.debug;
 
         return callback(null, queryData);
-    }
-}
-
-/**
- * Takes an array of contexts and returns geojson features, recording
- * used indexes along the way. Any of our `raw contexts => final output`
- * logic will ideally end up here.
- *
- * @param {Carmen} geocoder A carmen object used to generate the results
- * @param {array} contexts A contexts array of results to convert to features
- * @param {object} options A geocoder query options object
- * @return {object} Object with GeoJSON features on the 'features' property.
- */
-function toFeatures(geocoder, contexts, options) {
-    let features = [];
-    const indexes = {};
-    const by_place_name = {};
-
-    for (let i = 0; i < contexts.length; i++) {
-        const context = contexts[i];
-        const index = geocoder.indexes[context[0].properties['carmen:index']];
-        if (!filter.featureAllowed(index, context[0], options)) continue;
-
-        // Convert this context array to a feature
-        const feature = ops.toFeature(context, index.geocoder_format, options.language, options.languageMode, options.debug, geocoder, options.clipBBox, options.routing);
-
-        if (!options.allow_dupes) {
-            const types = typesFromIndex(index);
-
-            // Dedupe by place name, prefer feature with non-omitted, non-interpolated geom
-            const keys = [feature.place_name];
-            if (context[0].properties['carmen:spatialmatch'] !== undefined && context[0].properties['carmen:address'] !== undefined) {
-                keys.push(uniqueAddressId(geocoder, types, context));
-            }
-
-            const previous = keys.reduce((m, v) => (m || by_place_name[v] || m), false);
-            if (previous) {
-                if (features[previous.index].address && !feature.properties.address) {
-                    continue; // Don't allow street fallbacks to replace omitted or interpolated addresses
-                }
-                else if (features[previous.index].geometry.omitted && !feature.geometry.omitted) {
-                    features[previous.index] = feature;
-                }
-                else if (features[previous.index].geometry.interpolated && !feature.geometry.interpolated) {
-                    features[previous.index] = feature;
-                }
-                continue;
-            }
-
-            keys.forEach((k) => {
-                by_place_name[k] = { feature: feature, index: features.length };
-            });
-        }
-        features.push(feature);
-
-        // Record index usage for this feature
-        if (options.indexes) for (let k = 0; k < context.length; k++) {
-            indexes[context[k].properties['carmen:index']] = true;
-        }
-    }
-
-    // Re-sort array if we've changed composition
-    if (features.length !== contexts.length) {
-        features.sort((a, b) => b.relevance - a.relevance);
-    }
-
-    if (options.limit) features = features.slice(0, options.limit);
-
-    const data = { features: features };
-    if (options.indexes) data.indexes = Object.keys(indexes);
-    return data;
-}
-
-/**
- * Generate an unique Id for an address results based on the cover text and
- * the context array.
- *
- * @param {object} geocoder - carmen object
- * @param {boolean|Set} types - set of types
- * @param {Array<object>} context - context array
- * @returns {string} unique result id
- */
-function uniqueAddressId(geocoder, types, context) {
-    const coverText = new Map();
-    const covers = context[0].properties['carmen:spatialmatch'].covers;
-    let text = '';
-    for (let i = 0; i < covers.length; i++) {
-        text += ` ${covers[i].text}`;
-    }
-    coverText.set(context[0].properties['carmen:index'], text);
-
-    for (let j = 1; j < context.length; j++) {
-        const indexName = context[j].properties['carmen:index'];
-        // If we already have an string to represent this context, skip it.
-        if (coverText.has(indexName)) continue;
-        // If we got a list of acceptable types, respect it.
-        if (!types || types.has(geocoder.indexes[indexName].type)) {
-            coverText.set(indexName, context[j].properties['carmen:extid']);
-        }
-    }
-
-    return '_' + Array.from(coverText.values()).join(':');
-}
-
-/**
- * Build a list of index types from a format string
- * @param {object} index - carmen index
- * @returns {Set} set of index types
- */
-function typesFromIndex(index) {
-    if (index.geocoder_format && index.geocoder_format.default) {
-        const p = /(?:[^{]*)\{([^.]+)\.[^}]+\}/y;
-        const types = new Set();
-        let m;
-        // eslint-disable-next-line no-cond-assign
-        while (m = p.exec(index.geocoder_format.default)) types.add(m[1]);
-        return types;
-    }
-    else {
-        return false;
     }
 }
